requirements.txt
coloredlogs
discord.py
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-storage
